refactor: api_error.rb position for better architecture
